Version 3.44.05
Found and fixed several misguided signed-to-unsigned comparisons
Extended character set for XPMs to 92
Resaved unoptimal XPM icons
Fixed a couple bugs

+++ src/memory.c
@@ -318,7 +318,7 @@ wjmem *wjmemnew(int minsize, int incr)
 	wjmem *mem;
 
 	if (incr <= 0) incr = WJMEM_DEFINCR;
-	if (minsize <= sizeof(wjmem)) minsize = WJMEM_DEFSIZE;
+	if (minsize <= (int)sizeof(wjmem)) minsize = WJMEM_DEFSIZE;
 	mem = calloc(1, minsize);
 	if (mem)
 	{
